This is a note to the reader of this Git history
Commits were made very short, in order to be easir to read. In a real project, we would git-squash the commits in the PR, and rebase it on the newest version of the `dev` branch. That way we get a very clean git history

Commits in the "middle" of the branch included lots of prints for debugging, in order to ilustrate the development process